Make butons work
Changed the way buttons are created and given functionality.
Instead of a csv string now they are created with a dict,
this allows me to pass them the text, their values and the
object they will interact with.
So far I have only done it with stroked buttons because
I still dont have something for the menu buttons to
interact with.