MultiSession now working
Kind of, it is working alright, but not quite finished.

You can open several orders an switch between them quite nicely,
but it's far from finished.

Several major changes where implemented for this to work.

2 new kinds of buttons where created based heavily in the
previous buttons.

Features to be added:
- Link each button to it's order number not it's list index.
- Fix the overflow, so I either set a maximum number of orders
that can be open at any given time or I find a way to make it
scrollable.
- Change the style of the currently active order, so you
can know which one is active.