Fixed ActiveSession index bug
The previous way of choosing the active session was flawed.

If you deleted the active session it activated the first
session but it wouldn't show it as active.
And a few other weird things happened because I was using
the item index in the list to choose the current active session.
Now it works with the session Id so it is fixed.