OrderTotal partially working
It is working allright, but not finished yet, it just updates
the total but it is still lacking a lot of functionality.

Needed changes are:
- Adding tax calculation
- Adding discount calculation
- Styling

Changes to Buttons are completely unrelated but I left them
because I was going to do them later anyway, and they
don't add or remove any functionality.