Linux test passed
It seems the error was a sort of memory management issue,
but I dont know what it was.
It is solved for now, unless you select all products and/or click
buttons like a maniac.

If you do that it sort of crashes generating hundreds of weird
windows and you have to restart the software.

But it is such an unlikely scenario I wont bother fixing it now.