Extra data in OrderTotal
Now it shows all data to be shown, but it still lacks styling, and a way
to choose wether or not an invoice will be generated.

Also a way to add a discount, and for that here's a note:
If I want to add a specific disccount (eg. total minus
a $14 drink) it's easier to do it by calculating a discount that gives
that amount (I guess).