Editions for Ticket
Finally decided how to pass the data to the ticket, instead
of a long list I used a dict, so I changed collector to
pass the data as such.
Also added a way for the ticket to be called from the print
button, this is just as a test because that button is
meant to print the simplified version of the ticket.