Print on Raspberry Pi
Most vendors don't have drivers for ARM processors,
so the printer method was modified to add a direct method
of communication with compatible printers.

As the system can work both on pc's with the printer
drivers and without it will first try to print without
the drivers, if it is not setup, then it will fallback
to try print with the drivers, if this fails too
it will just do nothing.